manipulate_updated

+++ StartPage.py
@@ -11,6 +11,12 @@ page = st.sidebar.radio("Navigation", ["Home", "Manipulate Data", "Build Model",
 
 parent_data_set = None
 # Home page
+if "button" not in st.session_state:
+        st.session_state.button = {}
+
+def clicked(button):
+        st.session_state.button[button] = True
+
 if page == "Home":
     st.markdown('<p style="text-align: center;"><img src="Frontend\Data\logo-black.png" alt="Logo" style="border-radius: 50%; max-width: 200px;"></p>', unsafe_allow_html=True)
     st.title("Welcome to EasyML!")
@@ -18,15 +24,14 @@ if page == "Home":
     st.write("Explore the features in the sidebar.")
     # st.image("Frontend\Data\logo-black.png", use_column_width=True, output_format="png", channels="BGR", width=200, height=200, caption="EasyML")  # Replace with your logo
     
-    if "csv_clicked" not in st.session_state:
-        st.session_state.csv_clicked = {1:False}
-        st.session_state.clm_selected = {1 : False}
+    if "csv_upload" not in st.session_state.button:
+        st.session_state.button["csv_upload"] = False
+        
+    if "clicked" not in st.session_state.button:
+        st.session_state.button["clicked"] = False
 
-    def clicked(button):
-        st.session_state.csv_clicked[button] = True
-
-    st.button("Let's get started", on_click=clicked, args=[1])
-    if st.session_state.csv_clicked[1]:
+    st.button("Let's get started", on_click=clicked, args=["clicked"])
+    if st.session_state.button["clicked"]:
         st.header("You are now ready for your machine learning experience")
 
         # CSV upload button
@@ -38,10 +43,14 @@ if page == "Home":
         try:
             df = pd.read_csv(uploaded_file)
             st.write(df.head())
+            st.session_state.button["csv_upload"] = True
+            if "data_Set" not in st.session_state:
+                st.session_state.data_set = {"parent_Data_set" : df}
             parent_data_set = df
 
         except:
-            st.caption("Unable to Read CSV")
+            if not st.session_state.button["csv_upload"]:
+                st.caption("Unable to Read CSV")
 
     
 
@@ -49,13 +58,21 @@ if page == "Home":
 elif page == "Manipulate Data":
     st.title("Manipulate Data")
     st.write("Here you can process your data before selecting a training...")
-    manipulated_data = parent_data_set
+    if "data_set" in st.session_state:
+        st.session_state.data_set["manipulated_data"] = st.session_state.data_set["parent_Data_set"]
+    manipulated_data = st.session_state.data_set["manipulated_data"]
     selected_operation = st.selectbox("Select the operation", ["None", "drop column","impute"])
 
     if selected_operation == "drop column":
+        if "drop_button" not in st.session_state.button:
+            st.session_state.button["drop_button"] = False
         selected_columns = st.multiselect("select columns to drop",list(manipulated_data.columns))
-
-        data_drop_strategy = drop_column(selected_columns, manipulated_data)
+        st.button("drop selected columns", on_click=clicked, args= ["drop_button"])
+        if selected_columns and st.session_state.button["drop_button"]:
+            data_drop_strategy = drop_column(selected_columns, manipulated_data)
+            st.session_state.data_set["manipulated_data"] = data_drop_strategy
+            manipulated_data = st.session_state.data_set["manipulated_data"]
+            st.write(manipulated_data.head())
